pyproject.toml
[tool.poetry]
name = "team-up-test-task"
version = "0.1.0"
description = "IQ/EQ test API for TeamUP"
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = [{include = "team_up_test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-split-settings = "^1.2.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
uwsgi = "^2.0.21"
django-cors-headers = "^4.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
loguru = "^0.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
